# perf(transform): Improve AggregateToArray Throughput